+++ agents/builder/src/task/review_flow.py
@@ -27,7 +27,7 @@ if __name__ == "__main__":
 from anthropic.types import ToolUseBlock
 from src.get_file_list import get_file_list
 from src.task.setup import setup_client
-from src.task.retry_utils import (
+from src.utils.retry import (
     execute_tool_with_retry,
     send_message_with_retry,
 )
